requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.139
botocore==1.34.139
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nodeenv==1.9.1
openai==1.35.10
orjson==3.10.6
packaging==24.1
parameterized==0.9.0
prisma==0.13.1
psycopg2-binary==2.9.9
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tokenizers==0.19.1
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240622
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
